refactor: change the way heartbeat is computed
Define the delta as the distance between today (System.currentTimeMillis()) and the last received data from the client. This distance is in milliseconds, if too high the heartbeat will notify a connection lost. The use of a double instead of an integer makes possible to reset the delta something that previously would have been possible only by if the server's thread somehow stops which rarely happens.

+++ src/main/java/it/polimi/ingsw/network/heartbeat/HeartBeatMessage.java
@@ -4,41 +4,24 @@ import it.polimi.ingsw.network.NetworkMessage;
 import it.polimi.ingsw.network.Type;
 
 import java.io.Serializable;
-import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * HeartBeatMessage represents the message containing the sender and the id number of the message.
  */
 public class HeartBeatMessage extends NetworkMessage implements Serializable{
-    private final int id;
-    //private static final AtomicInteger incrementalId = new AtomicInteger();
-
-    //private static int getIncrementedId() {
-    //    return incrementalId.addAndGet(1);
-    //}
-
-    //public static int getIncrementalId() {
-    //    return incrementalId.get();
-    //}
-
-    // default constructor needed by gson
-  //  public HeartBeatMessage() {
-  //      super(Type.HEARTBEAT, "uninitializedSender");
-  //      //this.id = getIncrementedId();
-  //  }
-
+    private final double id;
     /**
      * Constructs a <code>HeartBeatMessage</code> with the <code>sender</code> and the <code>id</code> provided.
      *
      * @param sender the sender of the heartBeat.
      * @param id     the identification of the heartBeat message.
      */
-    public HeartBeatMessage(String sender, int id) {
+    public HeartBeatMessage(String sender, double id) {
         super(Type.HEARTBEAT, sender);
         this.id = id;
     }
 
-    public int getId() {
+    public double getId() {
         return id;
     }
 }
